John Davis: Fixed issues with animation sword missing (had swing animation on the wrong layer :P) added website code to source control.

+++ JDBaconNewUnity/Assets/Scripts/JDBaconUnityScripts/Classes/JDIWeapon.cs
@@ -12,6 +12,6 @@ public interface JDIWeapon : JDIObject
 {
     int DamageAmount { get; set; }
     bool IsActive { get; set; }
-    int CooldownTime { get; set; }
+    float CooldownTime { get; set; }
     float Range { get; set; }
 }
